// packages/formulize/src/key.enum.ts
/* eslint-disable @typescript-eslint/no-duplicate-enum-values */

/**
 * The keyboard event values from `e.key`.
 */
export enum KeyCodes {
  Backspace = "Backspace",
  Tab = "Tab",
  Shift = "Shift",
  Enter = "Enter",
  Control = "Control",
  Alt = "Alt",
  Meta = "Meta",
  Command = "Meta",
  Windows = "Meta",
  Pause = "Pause",
  CapsLock = "CapsLock",
  Escape = "Escape",
  Space = " ",
  PageUp = "PageUp",
  PageDown = "PageDown",
  End = "End",
  Home = "Home",
  LeftArrow = "ArrowLeft",
  UpArrow = "ArrowUp",
  RightArrow = "ArrowRight",
  DownArrow = "ArrowDown",
  Insert = "Insert",
  Delete = "Delete", // fn + Backspace on mac
  Zero = "0",
  ClosedParen = ")",
  One = "1",
  ExclamationMark = "!",
  Two = "2",
  AtSign = "@",
  Three = "3",
  Hash = "#",
  PoundSign = "#",
  Four = "4",
  DollarSign = "$",
  Five = "5",
  PercentSign = "%",
  Six = "6",
  Caret = "^",
  Hat = "^",
  Seven = "7",
  Ampersand = "&",
  Eight = "8",
  Asterik = "*",
  Nine = "9",
  OpenParen = "(",
  Plus = "+",

  // code 61
  Equals = "=",
  // all are code 187
  PlusMinus = "±",
  NotEqual = "≠",
  EqualsMultiply = "×",

  Subtract = "-",
  LongMinus = "–",

  Period = ".",
  Divide = "/",
  DivideAlt = "÷", // Option + / on mac
}